Refactored code out of compress.go into template
